requirements.txt
fastapi==0.97.0
aiofiles==23.2.1
sqlalchemy==2.0.19
pyjwt==2.8.0
passlib==1.7.4
aiosmtplib==2.0.2
jinja2==3.1.2
uvicorn==0.22.0
langchain==0.0.263
python-dotenv==1.0.0
huggingface_hub==0.15.1
pydantic==1.10.12
sqlalchemy-citext ==1.8.0
SQLAlchemy-Utils ==0.41.1
starlette ==0.27.0
anyio ==3.7.1
chainlit ==0.6.2
cryptography == 41.0.2
fastapi-mail == 1.3.1
python-decouple ==3.8
fastapi-sessions ==0.3.2
passlib==1.7.4
python-multipart==0.0.6
passlib==1.7.4
diffusers
transformers
accelerate
scipy
safetensors










